mm/huge_memory: cap max length on debugfs file entry folio split
Don't allow to query beyond a mapped file's length. Since this is just
a debugfs interface allow userspace to be lazy and use a large value so
we can just use the entire file.

Without this we can end up wasting cycles looking for folios which
just don't exist for no good reason.

Fixes: fa6c02315f745 ("mm: huge_memory: a new debugfs interface for splitting THP tests")
Reviewed-by: Zi Yan <ziy@nvidia.com>
Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>