selftests/mm: use global variable to not optimize mmap read variable
create_pagecache_thp_and_fd() in split_huge_page_test.c used the
variable dummy to perform mmap read.

However, this test was skipped even on XFS which has large folio
support. The issue was compiler (gcc 13.2.0) was optimizing out the
dummy variable, therefore, not creating huge page in the page cache.

Make it as a global variable to force the compiler not to optimize out
the loop where we read from the mmaped addr.

Signed-off-by: Pankaj Raghav <p.raghav@samsung.com>