module: return error from module_enable_rodata_ro_after_init()
The success path is not anymore success if
module_enable_rodata_ro_after_init() returns with error. This is
unlikely to happen since we check first if memory can be set readonly
first. In case it happens, we WARN_ONCE and continue as if it didn't
happpend returning 0 from do_init_module(). Ensure the success path
returns the error to let user space report that to the user.

Signed-off-by: Daniel Gomez <da.gomez@samsung.com>