nvme-pci: allow to configure nr segments
The maximum number of elements in a DMA scatter/gather list
(max_segments) is a hardcoded number in the NVMe driver introduced after
patch 943e942e6266f22babee5efeb00f8f672fbff5bd "nvme-pci: limit max IO
size and segments to avoid high order allocations". The max number has
been later optimized in commit 7846c1b5a5db8bb8475603069df7c7af034fd081
"nvme-pci: place descriptor addresses in iod" allowing 128 segments per
page instead of 127 that we had initially.

I/Os are therefore, limited to accomodate a max of 128 segments per page
per request. But this allocation is backed with a mempool, so requests
are always guaranteed.

This limitation however, makes the NVMe largest command to be 512 KiB
in size for a system with 4 KiB page size

    Dani:

    true? NVME_CTRL_PAGE_SIZE is another hardcoded value set to 4 KiB...
    can 128 segments send more than 512 KiB in 64 KiB systems?).

However, the maximum transfer driver size is specified to be 8192 KiB.

    Dani:

    64k * 128 = 8 MiB, is this number?
    or
    5 PRP list = 5 * 512 entries = 10 MiB

In order to reach the maximum tranfer size in the PRP path we need
to allocate more that 128 segments, so the pool tracking all these
descriptors needs to increase accordingly. So, ensure the pool is
as big as the user wants it to be.

Note that the pool will be bigger for all NVMe devices and will be as big as
requested for all cases. Even if the NVMe does not support it.

The pool won't be as big as the device Maximumm Data Transfer Size
(MDTS).

    Dani:

    Do we need to move BUILD_BUG_ON to something else?

Signed-off-by: Daniel Gomez <da.gomez@samsung.com>